pyproject.toml
[project]
name = "js-secret-santa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "ruff>=0.13.2",
]
